Bug fix in abundant_tax_physeq
ps_tbl was taken as the base for top n taxa and the given phyloseq object trimmed on the basis of that. This resulted in incorrect top n taxa being put out if the top 10 taxa in ps_tbl were not represented in physeq. This is now fixed by filtering ps_tbl to the taxa actually contained in physeq before operating on it. Generating ps_tbl fresh if the taxa in physeq are not equal to the taxa contained in ps_tbl generally will be considered